more accurate benchmark comparisons
Previous logic created a new list of randomized values for each test,
so that testing one B+Tree might insert or delete keys in some sort of
optimized order, while other B+Tree instances would have a different
list of randomized values that are in a worst case order for a
particular operation.

This caused benchmarks for the various B+Tree instances to not show
comparable numbers.

This change creates the list of randomized keys once for all B+Tree
instances, and uses that list for each of the benchmarks.