Comparable interface no longer requires Greater method
Rather than require both Less and Greater methods, requires Less
method, and code switches operand order when it needs to check for
when one operand is greater than the other operand.

Most of the time an Equal operation would really be required, but
equality is presumed true when !(a.Less(b) || b.Less(a)).